+++ platform/nucleo-u5a5zj/Adapters/FlexibleMemoryController/FlexibleMemoryController.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "IFlexibleMemoryController.hpp"
+#include "application/System/Ports/IFlexibleMemoryController.hpp"
 
 namespace ATC {
 class FlexibleMemoryController : public IFlexibleMemoryController {
